windows: Change wParam size to 64 bits
The wParam message parameter is used to send the callback object pointer
to WM_WIN_CALLBACK messages. As it is fixed as a 32 bit value in DALi,
when running on 64 bits systems it sends an invalid pointer to message
receiver.

+++ dali/internal/window-system/windows/platform-implement-win.h
@@ -39,7 +39,7 @@ namespace WindowsPlatformImplementation
 
 bool PostWinThreadMessage(
     _In_ uint32_t Msg,
-    _In_ uint32_t wParam,
+    _In_ uint64_t wParam,
     _In_ uint64_t lParam,
     _In_ uint64_t threadID = -1 );
 
@@ -83,7 +83,7 @@ public:
 
   bool PostWinMessage(
     _In_ uint32_t Msg,
-    _In_ uint32_t wParam,
+    _In_ uint64_t wParam,
     _In_ uint64_t lParam );
 
   int32_t GetEdgeWidth();
